user tests implemented

+++ src/main/java/com/cesarFrancisco/votePage/domain/services/UserService.java
@@ -4,6 +4,7 @@ import com.cesarFrancisco.votePage.api.dto.AuthenticationDto;
 import com.cesarFrancisco.votePage.configs.JwtUtils;
 import com.cesarFrancisco.votePage.domain.entities.User;
 import com.cesarFrancisco.votePage.domain.repositories.UserRepository;
+import com.cesarFrancisco.votePage.exceptions.NonAuthorizedException;
 import com.cesarFrancisco.votePage.exceptions.ObjectNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -56,6 +58,8 @@ public class UserService {
 
     public User insert(User user) {
 
+        user.setEmail(user.getEmail().toLowerCase().trim());
+
         user.setPassword(passwordEncoder.encode(user.getPassword()));
 
         user.setCreatedAt(new Date());
@@ -64,6 +68,21 @@ public class UserService {
 
     public User update(Long id, User user) {
         try {
+
+            String email = SecurityContextHolder.getContext().getAuthentication().getName();
+            Optional<User> requestUserOptional = userRepository.findByEmail(email);
+
+            if (requestUserOptional.isEmpty()) {
+                throw new ObjectNotFoundException("User not found");
+            }
+
+            User requestUser = requestUserOptional.get();
+
+            if (!requestUser.getId().equals(id)) {
+                throw new NonAuthorizedException("You can't update another user account");
+            }
+
+
             User oldUser = userRepository.getReferenceById(id);
             updateUser(oldUser, user);
             return userRepository.save(oldUser);
@@ -73,11 +92,30 @@ public class UserService {
     }
 
     private void updateUser(User oldUser, User newUser) {
-        oldUser.setName(newUser.getName());
-        oldUser.setEmail(newUser.getEmail());
+
+        if (newUser.getName() != null && !newUser.getName().isEmpty() && !newUser.getName().isBlank()) {
+            oldUser.setName(newUser.getName());
+        }
+        if (newUser.getEmail() != null && !newUser.getEmail().isEmpty() && !newUser.getEmail().isBlank()) {
+            oldUser.setEmail(newUser.getEmail().toLowerCase().trim());
+        }
     }
 
     public void delete(Long id) {
+
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<User> userOptional = userRepository.findByEmail(email);
+
+        if (userOptional.isEmpty()) {
+            throw new ObjectNotFoundException("User not found");
+        }
+
+        User user = userOptional.get();
+
+        if (!user.getId().equals(id)) {
+            throw new NonAuthorizedException("You can't delete another user account");
+        }
+
         userRepository.deleteById(id);
     }
